Remove substrate-archive-common
