+++ substrate-archive-backend/Cargo.toml
@@ -36,6 +36,3 @@ sp-storage = { git = "https://github.com/paritytech/substrate", branch = "master
 sp-trie = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-version = {  git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-wasm-interface = {  git = "https://github.com/paritytech/substrate", branch = "master" }
-
-[features]
-test_rocksdb = []
