+++ substrate-archive/Cargo.toml
@@ -7,9 +7,11 @@ edition = "2018"
 [dependencies]
 # external
 async-trait = "0.1"
+chrono = "0.4.19"
 coil = "0.2"
-log = "0.4"
+dirs = { version = "3", package = "directories" }
 fdlimit = "0.2"
+fern = { version = "0.6", features = ["colored"], optional = true }
 flume = "0.10.2"
 futures = "0.3"
 hashbrown = { version = "0.9", features = ["inline-more"] }
@@ -18,6 +20,7 @@ itertools = "0.10"
 itoa = "0.4.7"
 # Just a simple wrapper around std::thread that `joins on drop`
 jod-thread = "0.1.2"
+log = "0.4"
 num_cpus = "1"
 parking_lot = "0.11"
 rmp-serde = "0.15"
@@ -27,6 +30,8 @@ smol = "1.2"
 sqlx = { version = "0.5", default-features = false, features = ["postgres", "macros", "runtime-async-std-rustls", "migrate", "json", "offline", "chrono"] }
 tempfile = "3.2.0"
 thiserror = "1.0"
+tracing = "0.1.23"
+tracing-subscriber = "0.2.15"
 xtra = { version = "0.5.0-rc.1", features = ["with-smol-1"] }
 
 # Parity
@@ -47,10 +52,6 @@ sp-storage = { git ="https://github.com/paritytech/substrate", branch = "master"
 
 # Workspace
 substrate-archive-backend = { path = '../substrate-archive-backend' }
-substrate-archive-common = { path = '../substrate-archive-common' }
-tracing-subscriber = "0.2.15"
-tracing = "0.1.23"
-chrono = "0.4.19"
 
 [dev-dependencies]
 polkadot-service = { package = "polkadot-service", git = "https://github.com/paritytech/polkadot", branch = "master" }
@@ -64,5 +65,4 @@ dotenv = "0.15.0"
 
 [features]
 default = ["logging"]
-logging = ["substrate-archive-common/logging"]
-test_rocksdb = []
+logging = ["fern"]
