+++ substrate-archive/src/lib.rs
@@ -21,8 +21,9 @@
 pub use sc_executor::native_executor_instance;
 pub use sp_blockchain::Error as BlockchainError;
 pub use sp_runtime::MultiSignature;
+
 #[cfg(feature = "logging")]
-pub use substrate_archive_common::util::init_logger;
+pub mod logger;
 
 mod actors;
 pub mod archive;
@@ -31,6 +32,7 @@ mod error;
 mod migrations;
 mod tasks;
 mod types;
+mod util;
 mod wasm_tracing;
 
 pub use self::actors::System;
